requirements.txt
flask
flask-sqlalchemy
Flask-Login
Flask-JWT-Extended
flask-swagger-ui
Flask
Werkzeug
reportlab